refactor ansi tester project into multiple files for clarity.
