+++ projects/ansi/Tester/Fatal.go
@@ -1,0 +1,8 @@
+package ansi
+
+import "fmt"
+
+// Fatal - A Fatal error in a test
+func (test *T) Fatal(msg ...any) {
+	test.Failf(fmt.Sprintln(msg...))
+}
