+++ projects/ansi/Tester/Fatalf.go
@@ -1,0 +1,6 @@
+package ansi
+
+// Fatalf - A Fatal error in a test
+func (test *T) Fatalf(format string, msg ...any) {
+	test.Failf(format, msg...)
+}
