+++ projects/ansi/Tester/T.go
@@ -1,0 +1,11 @@
+package ansi
+
+import "testing"
+
+// T - The state of the Ansi Tester
+type T struct {
+	t    *testing.T
+	pass int
+	fail int
+	skip int
+}
